calls: move recording tests to own file
